feat: expand cli (#13)
* feat: add version command

Add the version command which prints the version of pypk

* feat: dynamically select default minimum python version

Dynamically select the default minimum python version based on the python version executing
pypk. For example, if one runs pypk in a 3.6.13 environment, the default minimum version will
be 3.6.0. Likewise, if one runs pypk in a 3.7.3 environment, the default minimum version will
be 3.7.0.

* feat: enable git initialization to be toggled

* fix: properly handle packages with dashes in their name

Properly handle packages with dashes in their name, replacing them with underscores
when creating the src directory.

* feat: perform minor validation of the python version

* refactor: replace -v/--version flag with -p/--py-version

* feat: expand setup.py template

Expand setup.py template to include long_description and install_requires. Additionally, let
description be specified via the cli.

* feat: enable a top-level tests directory to be created

* feat: enable config values to be retrieved/set via cli

* feat: enable config support for init_git and tests_dir